requirements.txt
torch>=2.2.0
transformers>=4.45.0
peft>=0.12.0
datasets>=2.20.0
accelerate>=0.30.0
mlx>=0.18.0
mlx-lm>=0.18.0
wandb>=0.17.0
numpy>=1.26.0
tqdm>=4.66.0
PyYAML>=6.0
